Correção da listagem de pedidos + remoção de parametro id em urls para itens privados

+++ api/src/main/java/com/commerce/api/service/LojaService.java
@@ -7,6 +7,7 @@ import com.commerce.api.model.Produto;
 import com.commerce.api.model.Usuario;
 import com.commerce.api.model.dto.LojaDTO;
 import com.commerce.api.model.dto.LojaUpdateDTO;
+import com.commerce.api.model.dto.RequestDTO;
 import com.commerce.api.repository.LojaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -31,7 +32,7 @@ public class LojaService {
     private ProdutoService produtoService;
 
     public Usuario createNewLojaAccount(Loja newLoja) {
-        System.out.println("createNewLojaAccount");
+
         var saved = lojaRepository.save(newLoja);
         saved.add(linkTo(methodOn(LojaController.class).getById(saved.getId())).withSelfRel());
         return saved;
@@ -43,7 +44,7 @@ public class LojaService {
         loja.setPassword(encriptedPassword);
         Loja saved = lojaRepository.save(loja);
         saved.add(linkTo(methodOn(LojaController.class).getById(saved.getId())).withSelfRel());
-        System.out.println("createLoja");
+
         return saved;
     }
 
@@ -54,63 +55,37 @@ public class LojaService {
     public PagedModel<EntityModel<Loja>> getAllLojas(Pageable pageable) {
         Page<Loja> lojas = lojaRepository.findAll(pageable);
         lojas.forEach(c -> c.add(linkTo(methodOn(LojaController.class).getById(c.getId())).withSelfRel()));
-        System.out.println("getAllLojas");
+
         return assembler.toModel(lojas);
 
     }
 
     public Loja getLojaById(Long id) throws ResourceNotFoundException {
-        Loja loja = lojaRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Loja não encontrada."));
+        Loja loja = lojaRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Loja não encontrada."));
         loja.add(linkTo(methodOn(LojaController.class).getAll(0, 0, "asc")).withRel("Listagem"));
-        System.out.println("getLojaById");
+
         return loja;
     }
-
 
     public Loja updateLoja(String username, LojaUpdateDTO dto) throws ResourceNotFoundException {
         Loja loja = lojaRepository.findByUsername(username);
         Loja updated = lojaRepository.save(updateProperties(loja, dto));
         updated.add(linkTo(methodOn(LojaController.class).getById(updated.getId())).withSelfRel());
-        System.out.println("updateLoja");
+
         return updated;
     }
 
-    public void deleteLoja(Long id) {
-        System.out.println("deleteLoja");
-        try {
-            Loja loja = lojaRepository.findById(id).get();
+    public void deleteLoja(String username) {
+        if (lojaRepository.existsByUsername(username)){
+            Loja loja = lojaRepository.findByUsername(username);
             lojaRepository.delete(loja);
-        } catch (Exception e) {
-            System.err.printf("DELETE:Loja: %d não encontrado%n", id);
         }
     }
 
     public PagedModel<EntityModel<Produto>> getAllProdutos(String username) {
-        System.out.println("getAllProdutos");
+
         return produtoService.getAllProdutosByLoja(lojaRepository.findByUsername(username));
-    }
-
-    public void adicionarProduto(String username, Long produtoId) throws ResourceNotFoundException {
-        System.out.println("adicionarProduto");
-        Loja loja = lojaRepository.findByUsername(username);
-        Produto produto = produtoService.getProdutoById(produtoId);
-
-        produto.setLoja(loja);
-        loja.adicionarProduto(produto);
-
-        produtoService.save(produto);
-        lojaRepository.save(loja);
-
-    }
-
-    public void removerProduto(String username, Long produtoId) throws ResourceNotFoundException {
-        System.out.println("removerProduto");
-        Loja loja = lojaRepository.findByUsername(username);
-        Produto produto = produtoService.getProdutoById(produtoId);
-        produto.setLoja(null);
-        loja.removerProduto(produto);
-        produtoService.deleteProduto(username, produtoId);
-        lojaRepository.save(loja);
     }
 
     void save(Loja loja) {
@@ -118,7 +93,7 @@ public class LojaService {
     }
 
     private Loja updateProperties(Loja loja, LojaUpdateDTO dto) {
-        System.out.println("updateProperties");
+
         loja.setNome(dto.nome() != null ? dto.nome() : loja.getNome());
         loja.setCNPJ(dto.CNPJ() != null ? dto.CNPJ() : loja.getCNPJ());
         loja.setEmail(dto.email() != null ? dto.email() : loja.getEmail());
